Noise Refactor
I did not like the previous implementation, to play a contious noise it had to be called constantly which is a fucking stupid design choice by me.

To make it more robust and flexible, noises are now defined by an interface. This means the noise manager treats all noises the same, but each concrete noise can have different logic.

Implemented 2 noise types
- Single noise. This plays a single noise and begins decreasing immediately (useful for 1 time actions like a jump)
- Toggle noise. This allows you to toggle a noises playing state on or off. Note: this does not stop the noise immediately it still fades out based on the noise settings delay/duration (useful for actions such as moving where you are moving or not)