Initial Super Basic Interaction Setup
Just to see stuff I created a super simple interaction.
The player has an interactor script that can trigger with IInteractables.
You can then call the interact on that interactable, which at the moment is just logging out who is the interactor

The way it is set up it should also work with the monster as well as there is no input in the interactor script