Context Actions
Context actions are scriptable objects that can define behaviour, the interactables have a list of context actions that execute when the interactable is interacted with.
This allows quite a lot of flexability with what each interactable can do.

Bugfix -> updated the player to rotate to face the camera forward as the interactor hitbox would not rotate and therefore you could only interact in 1 direction