Objective Holder and Cloning Objectives
The objective holder class is so objectives can exist in the world without the need of interaction, but still needing scene references.

Cloning the objective data to the monobehaviour then setting the objective scriptable objects objective data to the one from the monobehaviour as the type mismatch was causing issues and this seems to work.

Added a test task giver so I could see the tasks active and check if there are issues