Depth Overhaul
Changed the depth system to work similar to the noise system. Have a list of IDepth, each frame minus the depth contribution.

A new depth consequence that increases the current depth over a duration when a task is failed.

To compensate for this I clamped the current depth value between 0 and the starting depth as I don't think its fair to be able to go below the starting depth