Duplicate Task Bugfix
It was possible to get the same task multiple times which is unintended, fixed it.

Well it can throw an error if there are no possible tasks as the player has them all. But this will basically be impossible to happen as the engine task prevents gaining other tasks as the sub does not rise